# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WAP — WordPress AI Platform
  version: 0.1.0
paths:
...
  /api/v1/auth/session:
    post:
      summary: Create Session
      operationId: create_session_api_v1_auth_session_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Session Api V1 Auth Session Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Revoke Session
      operationId: revoke_session_api_v1_auth_session_delete
      responses:
        '204':
          description: Successful Response
      security:
      - HTTPBearer: []
...
  /admin/login:
    get:
      summary: Get Login
      operationId: get_login_admin_login_get
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Post Login
      operationId: post_login_admin_login_post
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_post_login_admin_login_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/logout:
    get:
      summary: Logout
      operationId: logout_admin_logout_get
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/ui/agents:
    get:
      summary: Agents Page
      operationId: agents_page_admin_ui_agents_get
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/ui/agents/new:
    get:
      summary: Agents New Form
      operationId: agents_new_form_admin_ui_agents_new_get
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/ui/agents/{agent_id}/edit:
    get:
      summary: Agents Edit Form
      operationId: agents_edit_form_admin_ui_agents__agent_id__edit_get
      parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          title: Agent Id
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/ui/roles:
    get:
      summary: Roles Page
      operationId: roles_page_admin_ui_roles_get
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/ui/snippets:
    get:
      summary: Snippets Page
      operationId: snippets_page_admin_ui_snippets_get
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    AuthBody:
      properties:
        product:
          type: string
          minLength: 1
          title: Product
        license_key:
          type: string
          title: License Key
          default: ''
        site_url:
          type: string
          minLength: 1
          title: Site Url
        mode:
          type: string
          pattern: ^(product|orchestrator)$
          title: Mode
          default: product
        mcp_endpoint:
          type: string
          minLength: 1
          title: Mcp Endpoint
        wp_username:
          type: string
          minLength: 1
          title: Wp Username
        wp_app_password:
          type: string
          minLength: 1
          title: Wp App Password
        available_products:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Available Products
      type: object
      required:
      - product
      - site_url
      - mcp_endpoint
      - wp_username
      - wp_app_password
      title: AuthBody
    Body_post_login_admin_login_post:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: Body_post_login_admin_login_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer